# PERF: optimize a hot path in `Angle.__new__`