@neutrinoceros 2024-02-22
FWIW I'm game to try and speed up unit comparison to. I'll add it to my list :)