@pllim 2024-03-18
Applied but please check the log carefully even if it is green because I have not merged this:

* https://github.com/astropy/astropy/pull/16207