@mhvk 2024-03-18
Nice to see the benchmarks in action!

@eerovaher - I think @neutrinoceros addressed your comments and now the benchmarks work. Could you check that your change request can be dismissed (and this merged)?